+++ guillotina_rediscache/tests/test_cache.py
@@ -1,22 +1,25 @@
-from guillotina_rediscache.cache_strategy import RedisCache
-from guillotina.tests import mocks
-from guillotina_rediscache import cache
-from guillotina.tests.utils import create_content
 from guillotina import app_settings
-from guillotina_rediscache.interfaces import IRedisChannelUtility
 from guillotina.component import getUtility
+from guillotina.tests import mocks
+from guillotina.tests.utils import create_content
+from guillotina_rediscache import cache
+from guillotina_rediscache import serialize
+from guillotina_rediscache.cache_strategy import RedisCache
+from guillotina_rediscache.interfaces import IRedisChannelUtility
+
 import asyncio
 
 
 async def test_cache_set(redis, dummy_guillotina, loop):
     await cache.close_redis_pool()
     trns = mocks.MockTransaction(mocks.MockTransactionManager())
+    trns.added = trns.deleted = {}
     rcache = RedisCache(mocks.MockStorage(), trns, loop=loop)
     await rcache.clear()
 
     await rcache.set('bar', oid='foo')
     # make sure it is in redis
-    assert await rcache._conn.get('foo') == b'"bar"'
+    assert serialize.loads(await rcache._conn.get('foo')) == "bar"
     # but also in memory
     assert rcache._memory_cache.get('foo') == 'bar'
     # and api matches..
@@ -28,12 +31,13 @@ async def test_cache_set(redis, dummy_guillotina, loop):
 async def test_cache_delete(redis, dummy_guillotina, loop):
     await cache.close_redis_pool()
     trns = mocks.MockTransaction(mocks.MockTransactionManager())
+    trns.added = trns.deleted = {}
     rcache = RedisCache(mocks.MockStorage(), trns, loop=loop)
     await rcache.clear()
 
     await rcache.set('bar', oid='foo')
     # make sure it is in redis
-    assert await rcache._conn.get('foo') == b'"bar"'
+    assert serialize.loads(await rcache._conn.get('foo')) == "bar"
     assert rcache._memory_cache.get('foo') == 'bar'
     assert await rcache.get(oid='foo') == 'bar'
 
@@ -47,12 +51,13 @@ async def test_cache_delete(redis, dummy_guillotina, loop):
 async def test_cache_clear(redis, dummy_guillotina, loop):
     await cache.close_redis_pool()
     trns = mocks.MockTransaction(mocks.MockTransactionManager())
+    trns.added = trns.deleted = {}
     rcache = RedisCache(mocks.MockStorage(), trns, loop=loop)
     await rcache.clear()
 
     await rcache.set('bar', oid='foo')
     # make sure it is in redis
-    assert await rcache._conn.get('foo') == b'"bar"'
+    assert serialize.loads(await rcache._conn.get('foo')) == "bar"
     assert rcache._memory_cache.get('foo') == 'bar'
     assert await rcache.get(oid='foo') == 'bar'
 
@@ -65,13 +70,14 @@ async def test_cache_clear(redis, dummy_guillotina, loop):
 async def test_invalidate_object(redis, dummy_guillotina, loop):
     await cache.close_redis_pool()
     trns = mocks.MockTransaction(mocks.MockTransactionManager())
+    trns.added = trns.deleted = {}
     content = create_content()
     trns.modified = {content._p_oid: content}
     rcache = RedisCache(mocks.MockStorage(), trns, loop=loop)
     await rcache.clear()
 
     await rcache.set('foobar', oid=content._p_oid)
-    assert await rcache._conn.get(content._p_oid) == b'"foobar"'
+    assert serialize.loads(await rcache._conn.get(content._p_oid)) == "foobar"
     assert rcache._memory_cache.get(content._p_oid) == 'foobar'
     assert await rcache.get(oid=content._p_oid) == 'foobar'
 
@@ -84,13 +90,14 @@ async def test_invalidate_object(redis, dummy_guillotina, loop):
 async def test_subscriber_invalidates(redis, dummy_guillotina, loop):
     await cache.close_redis_pool()
     trns = mocks.MockTransaction(mocks.MockTransactionManager())
+    trns.added = trns.deleted = {}
     content = create_content()
     trns.modified = {content._p_oid: content}
     rcache = RedisCache(mocks.MockStorage(), trns, loop=loop)
     await rcache.clear()
 
     await rcache.set('foobar', oid=content._p_oid)
-    assert await rcache._conn.get(content._p_oid) == b'"foobar"'
+    assert serialize.loads(await rcache._conn.get(content._p_oid)) == "foobar"
     assert rcache._memory_cache.get(content._p_oid) == 'foobar'
     assert await rcache.get(oid=content._p_oid) == 'foobar'
 
@@ -109,13 +116,14 @@ async def test_subscriber_invalidates(redis, dummy_guillotina, loop):
 async def test_subscriber_ignores_trsn_on_invalidate(redis, dummy_guillotina, loop):
     await cache.close_redis_pool()
     trns = mocks.MockTransaction(mocks.MockTransactionManager())
+    trns.added = trns.deleted = {}
     content = create_content()
     trns.modified = {content._p_oid: content}
     rcache = RedisCache(mocks.MockStorage(), trns, loop=loop)
     await rcache.clear()
 
     await rcache.set('foobar', oid=content._p_oid)
-    assert await rcache._conn.get(content._p_oid) == b'"foobar"'
+    assert serialize.loads(await rcache._conn.get(content._p_oid)) == "foobar"
     assert rcache._memory_cache.get(content._p_oid) == 'foobar'
     assert await rcache.get(oid=content._p_oid) == 'foobar'